refactor: `useCreateSensorForm`
